Rename FolderProvider to FolderService and FolderService to FolderStateService
