+++ apps/browser/src/popup/services/services.module.ts
@@ -20,7 +20,7 @@ import { EventService } from "@bitwarden/common/abstractions/event.service";
 import { ExportService } from "@bitwarden/common/abstractions/export.service";
 import { FileDownloadService } from "@bitwarden/common/abstractions/fileDownload/fileDownload.service";
 import { FileUploadService } from "@bitwarden/common/abstractions/fileUpload.service";
-import { FolderService } from "@bitwarden/common/abstractions/folder/folder.service";
+import { FolderStateService } from "@bitwarden/common/abstractions/folder/folder-state.service.abstraction";
 import { I18nService } from "@bitwarden/common/abstractions/i18n.service";
 import { KeyConnectorService } from "@bitwarden/common/abstractions/keyConnector.service";
 import { LogService as LogServiceAbstraction } from "@bitwarden/common/abstractions/log.service";
@@ -148,7 +148,11 @@ function getBgService<T>(service: keyof MainBackground) {
       useFactory: getBgService<CryptoFunctionService>("cryptoFunctionService"),
       deps: [],
     },
-    { provide: FolderService, useFactory: getBgService<FolderService>("folderService"), deps: [] },
+    {
+      provide: FolderStateService,
+      useFactory: getBgService<FolderStateService>("folderStateService"),
+      deps: [],
+    },
     {
       provide: CollectionService,
       useFactory: getBgService<CollectionService>("collectionService"),
@@ -237,7 +241,7 @@ function getBgService<T>(service: keyof MainBackground) {
         return new VaultFilterService(
           getBgService<StateServiceAbstraction>("stateService")(),
           getBgService<OrganizationService>("organizationService")(),
-          getBgService<FolderService>("folderService")(),
+          getBgService<FolderStateService>("folderStateService")(),
           getBgService<CipherService>("cipherService")(),
           getBgService<CollectionService>("collectionService")(),
           getBgService<PolicyService>("policyService")()
