+++ apps/cli/src/commands/list.command.ts
@@ -1,7 +1,7 @@
 import { ApiService } from "@bitwarden/common/abstractions/api.service";
 import { CipherService } from "@bitwarden/common/abstractions/cipher.service";
 import { CollectionService } from "@bitwarden/common/abstractions/collection.service";
-import { FolderService } from "@bitwarden/common/abstractions/folder/folder.service";
+import { FolderStateService } from "@bitwarden/common/abstractions/folder/folder-state.service.abstraction";
 import { OrganizationService } from "@bitwarden/common/abstractions/organization.service";
 import { SearchService } from "@bitwarden/common/abstractions/search.service";
 import { Utils } from "@bitwarden/common/misc/utils";
@@ -26,7 +26,7 @@ import { CliUtils } from "../utils";
 export class ListCommand {
   constructor(
     private cipherService: CipherService,
-    private folderService: FolderService,
+    private folderStateService: FolderStateService,
     private collectionService: CollectionService,
     private organizationService: OrganizationService,
     private searchService: SearchService,
