+++ apps/web/src/app/settings/emergency-add-edit.component.ts
@@ -4,7 +4,7 @@ import { AuditService } from "@bitwarden/common/abstractions/audit.service";
 import { CipherService } from "@bitwarden/common/abstractions/cipher.service";
 import { CollectionService } from "@bitwarden/common/abstractions/collection.service";
 import { EventService } from "@bitwarden/common/abstractions/event.service";
-import { FolderService } from "@bitwarden/common/abstractions/folder/folder.service";
+import { FolderStateService } from "@bitwarden/common/abstractions/folder/folder-state.service.abstraction";
 import { I18nService } from "@bitwarden/common/abstractions/i18n.service";
 import { LogService } from "@bitwarden/common/abstractions/log.service";
 import { MessagingService } from "@bitwarden/common/abstractions/messaging.service";
@@ -29,7 +29,7 @@ export class EmergencyAddEditComponent extends BaseAddEditComponent {
 
   constructor(
     cipherService: CipherService,
-    folderService: FolderService,
+    folderStateService: FolderStateService,
     i18nService: I18nService,
     platformUtilsService: PlatformUtilsService,
     auditService: AuditService,
@@ -46,7 +46,7 @@ export class EmergencyAddEditComponent extends BaseAddEditComponent {
   ) {
     super(
       cipherService,
-      folderService,
+      folderStateService,
       i18nService,
       platformUtilsService,
       auditService,
