+++ apps/web/src/app/vault/folder-add-edit.component.ts
@@ -1,8 +1,8 @@
 import { Component } from "@angular/core";
 
 import { FolderAddEditComponent as BaseFolderAddEditComponent } from "@bitwarden/angular/components/folder-add-edit.component";
-import { FolderProviderAbstraction } from "@bitwarden/common/abstractions/folder/folder.provider.abstraction";
-import { FolderService } from "@bitwarden/common/abstractions/folder/folder.service";
+import { FolderStateService } from "@bitwarden/common/abstractions/folder/folder-state.service.abstraction";
+import { FolderServiceAbstraction } from "@bitwarden/common/abstractions/folder/folder.service.abstraction";
 import { I18nService } from "@bitwarden/common/abstractions/i18n.service";
 import { LogService } from "@bitwarden/common/abstractions/log.service";
 import { PlatformUtilsService } from "@bitwarden/common/abstractions/platformUtils.service";
@@ -13,12 +13,12 @@ import { PlatformUtilsService } from "@bitwarden/common/abstractions/platformUti
 })
 export class FolderAddEditComponent extends BaseFolderAddEditComponent {
   constructor(
-    folderService: FolderService,
-    folderProvider: FolderProviderAbstraction,
+    folderStateService: FolderStateService,
+    folderService: FolderServiceAbstraction,
     i18nService: I18nService,
     platformUtilsService: PlatformUtilsService,
     logService: LogService
   ) {
-    super(folderService, folderProvider, i18nService, platformUtilsService, logService);
+    super(folderStateService, folderService, i18nService, platformUtilsService, logService);
   }
 }
