+++ libs/common/src/abstractions/folder/folder.service.abstraction.ts
@@ -1,6 +1,6 @@
 import { Folder } from "@bitwarden/common/models/domain/folder";
 
-export class FolderProviderAbstraction {
+export class FolderServiceAbstraction {
   save: (folder: Folder) => Promise<any>;
   delete: (id: string) => Promise<any>;
 }
